Added P,I,D,F, and RPM constants, created state machine for cargo.
Added constants.
Initialized shooterPIDController which is a CANSparkMaxPIDController for the shooter motor. Set P,I,D,FF values for the PID controller based on new constants added. Created OperateShooter method to set the RPM of the shooter motor to the specified value in constants. Operated and stopped motors in OperateCargo based on state.